% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexwords-package.R
\docType{package}
\name{hexwords-package}
\alias{hexwords}
\alias{hexwords-package}
\title{hexwords: Words That Are Colours}
\description{
Provides a short list of words that are colours, currently only for English.
}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

Other contributors:
\itemize{
  \item Colin Fay [contributor]
}

}
\keyword{internal}
